requirements.txt
matplotlib==3.7.3
mpmath==1.3.0
numpy==1.24.4
pytorch_metric_learning==1.6.0
scikit_learn==1.1.1
scipy==1.8.1
setuptools==44.0.0
torch==2.4.1
tqdm==4.64.0
umap_learn==0.5.5
wandb==0.19.7